AME-1271 -- Non-Numeric Comparisons Completed. Finishing Tests cases for each Argument in this group and changes where applicable.
git-svn-id: https://svn.forgerock.org/openam/branches/openam_10.2.0_xacml3_JAS@4771 0f4defcf-c51a-4c67-9f44-6fb5eba73c5d

+++ openam/openam-entitlements/src/main/java/org/forgerock/openam/xacml/v3/Functions/DaytimedurationEqual.java
@@ -54,7 +54,7 @@ public class DaytimedurationEqual extends XACMLFunction {
         Long duration1 = getArg(0).asDayTimeDuration(pip);
         Long duration2 = getArg(1).asDayTimeDuration(pip);
 
-        if (duration1 == duration2) {
+        if (duration1.longValue() == duration2.longValue()) {
             retVal = FunctionArgument.trueObject;
         }
         return retVal;
